refactor: streamline Tripletex resource imports and remove setup.py file
