+++ tests/integration/test_tripletex.py
@@ -2,8 +2,12 @@ from datetime import datetime, timezone
 
 import pytest
 
-from .tripletex_resources.models import Country, CountryResponse
-from .tripletex_resources.setup import TripletexAPI, TripletexTestConfig
+from .tripletex_resources import (
+    Country,
+    CountryResponse,
+    TripletexAPI,
+    TripletexTestConfig,
+)
 
 
 @pytest.fixture
@@ -25,7 +29,7 @@ def test_api_configuration(api):
     assert api.client.config.auth_strategy is not None
 
     # Check that the auth strategy is a TripletexAuthStrategy
-    from .tripletex_resources.setup import TripletexAuthStrategy
+    from .tripletex_resources import TripletexAuthStrategy
 
     assert isinstance(api.client.config.auth_strategy, TripletexAuthStrategy)
