+++ tests/integration/tripletex_resources/__init__.py
@@ -3,15 +3,35 @@ from .auth import TripletexAuthStrategy
 from .client import TripletexClient
 from .config import TripletexConfig, TripletexTestConfig
 from .crud import TripletexCrud
-from .endpoints import TripletexCountries, TripletexSuppliers
+from .endpoints import TripletexCompany, TripletexCountries, TripletexSuppliers
+from .models import (
+    Company,
+    CompanyResponse,
+    Country,
+    CountryResponse,
+    Supplier,
+    SupplierResponse,
+    TokenSessionResponse,
+)
 
 __all__ = [
+    # Core components
     "TripletexAuthStrategy",
     "TripletexAPI",
     "TripletexClient",
     "TripletexConfig",
     "TripletexTestConfig",
     "TripletexCrud",
+    # Endpoints
+    "TripletexCompany",
     "TripletexCountries",
     "TripletexSuppliers",
+    # Models
+    "Company",
+    "CompanyResponse",
+    "Country",
+    "CountryResponse",
+    "Supplier",
+    "SupplierResponse",
+    "TokenSessionResponse",
 ]
